make batchrun more robust.
results overall are pretty successful on non face objects. the largest issue i have with current results is the fact that long sliver triangles can appear with multiple colors across it. it's like the optimization wanted to squash these triangles away, but the area penalty forces them to continue existing. these are also such small triangles that the edge split gain for them is too small to be worth it. it's a visually large glitch but an energetically small issue.

maybe i can detect when gradient flow and area penalty are countering each other on very slivered triangles and collapse those?
Alternatives also include removing area penalty altogether and frequently re-mesh.

another issue is that the area penalty keeps increasing size of the edge split triangles. The kind of morally right way to fix this is to add area penalty score to the edge subdivision score, and also have area penalty be weighted by saliency.

last issue is that the linear gradient flow is not saliency aware. edge split relies on degree 0 energy only. should probably fix this for completeness, but the linear case is meh so far. though that's a chicken or the egg kind of issue.